znr: change unsigned long to unsigned long long
Currently unsigned long type is used for sectors, on ILP32 this is 4
bytes and 8 bytes on LP64 [1]. As is, Zonar would be broken on 32bit
systems where for example, we could have sector ranges overflow.

[1] https://www.ibm.com/docs/en/ent-metalc-zos/3.1.0?topic=environments-ilp32-lp64-data-models-data-type-sizes

+++ src/znr_bg.h
@@ -10,22 +10,22 @@
  */
 struct znr_blockgroup {
         /* Starting sector */
-	unsigned long sector;
+	unsigned long long sector;
 
 	/* Number of sectors */
-	unsigned long nr_sectors;
+	unsigned long long nr_sectors;
 
 	/*
 	 * The device indicated write pointer sector offset within this
 	 * blockgroup. Valid if ZNR_BG_HAS_DEV_ZONE_WP is set in flags.
 	 */
-	unsigned long dev_zone_wp;
+	unsigned long long dev_zone_wp;
 
 	/*
 	 * The current write pointer offset within this blockgroup used by the
 	 * FS to issue writes. Valif if ZNR_BG_HAS_FS_WP is set in flags.
 	 */
-	unsigned long fs_wp;
+	unsigned long long fs_wp;
 
 	/* Is true if the backing zones has been mapped to this blockgroup */
 	bool has_zone_mapping;
@@ -40,7 +40,7 @@ struct znr_blockgroup {
 
 	/* Zones in this block group */
 	struct blk_zone **zones;
-	unsigned long nr_zones;
+	unsigned long long nr_zones;
 };
 
 #define ZNR_BG_SHARED_FLAGS ((1U << 24) - 1)
